# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Identity Onboarding Lab — Public API
  version: 1.0.0
  description: Public HTTP API contract for Sri Lankan investment customer onboarding and reviewer staff operations.
paths:
  /auth/otp/request:
    post:
      summary: Request contact-bound OTP challenge
      operationId: requestOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contactLookup]
              properties:
                contactLookup:
                  type: string
                  example: "+94771234567"
      responses:
        '200':
          description: OTP challenge issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  challengeId:
                    type: string
                    format: uuid
                  expiresInSeconds:
                    type: integer
                    example: 300
                  demoCodeHint:
                    type: string
                    nullable: true
                    description: Populated only in local demo mode when test provider is active
        '429':
          description: Rate limit exceeded
  /auth/otp/verify:
    post:
      summary: Verify OTP and create applicant session
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, otpCode]
              properties:
                challengeId:
                  type: string
                  format: uuid
                otpCode:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  userId:
                    type: string
                    format: uuid
                  token:
                    type: string
                  role:
                    type: string
                    example: "APPLICANT"
  /auth/staff/login:
    post:
      summary: Staff reviewer and admin authentication
      operationId: staffLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authenticated staff session
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  userId:
                    type: string
                    format: uuid
                  token:
                    type: string
                  role:
                    type: string
                    enum: [REVIEWER, ADMIN, AUDITOR]
...
  /applications/{id}:
    get:
      summary: Fetch application state and next action
      operationId: getApplication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
...
  /applications/{id}/biometric/challenges:
    post:
      summary: Request single-use cryptographically random camera challenge
      operationId: requestBiometricChallenge
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Challenge issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
  /applications/{id}/biometric/attempts:
    post:
      summary: Submit recorded camera challenge video
      operationId: submitBiometricAttempt
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [videoFile, challengeId, nonce, expectedVersion]
              properties:
                videoFile:
                  type: string
                  format: binary
                challengeId:
                  type: string
                  format: uuid
                nonce:
                  type: string
                expectedVersion:
                  type: integer
      responses:
        '202':
          description: Video accepted for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  attemptId:
                    type: string
                    format: uuid
  /applications/{id}/checks:
    get:
      summary: Retrieve all verification checks and capability labels
      operationId: getChecks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Verification checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VerificationCheck'
...
components:
  schemas:
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lifecycle:
          type: string
          enum: [DRAFT, PROCESSING, READY_TO_SUBMIT, SUBMITTED, UNDER_REVIEW, COMPLETED, WITHDRAWN, REJECTED]
        decision:
          type: string
          enum: [PENDING, RECAPTURE_REQUIRED, REVIEW_REQUIRED, DEMO_CHECKS_PASSED, APPROVED, REJECTED]
        operatingMode:
          type: string
          enum: [LOCAL_DEMO, INSTITUTION]
        version:
          type: integer
        currentDocumentGeneration:
          type: integer
        currentBiometricGeneration:
          type: integer
        submittedAt:
          type: string
          format: date-time
          nullable: true
    ChallengeResponse:
      type: object
      properties:
        challengeId:
          type: string
          format: uuid
        nonce:
          type: string
        steps:
          type: array
          items:
            type: string
            enum: [TURN_LEFT, TURN_RIGHT, BLINK]
        expiresAt:
          type: string
          format: date-time
...
    VerificationCheck:
      type: object
      properties:
        id:
          type: string
          format: uuid
        checkType:
          type: string
        status:
          type: string
          enum: [NOT_RUN, PENDING, RUNNING, PASS, FAIL, INCONCLUSIVE, ERROR, UNAVAILABLE, UNSUPPORTED]
        executionMode:
          type: string
          enum: [REAL, REAL_EXPERIMENTAL, MOCKED, UNAVAILABLE]
        score:
          type: number
          nullable: true
        scoreType:
          type: string
          nullable: true
        threshold:
          type: number
          nullable: true
        provider:
          type: string
        modelVersion:
          type: string
        reasonCodes:
          type: array
          items:
            type: string